Create Generic Attackable Object and Unit
Create a attackable class which will be used by both units and buildings to store their health, what happens when they are attacked, name and a description about it.
Create a sub class of attackable which will be used to store a unit’s damage information and other unit specific information.
Create a test unit sub class of unit that just test that the unit correctly executes its die function when it has received damage to reduce its health below 0.